requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==6.2.0
#psycopg2==2.7.4 --no-binary psycopg2

# Django
django

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework
django-filter==2.0.0

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

# Celery
tornado==5.1.1
redis==3.2.0
django-redis==4.10.0
celery==4.2.1
flower==0.9.2
